Add 9.0nann1 update

+++ LolexTools.py
@@ -25,7 +25,7 @@ except(ImportError):
     else:
         os.system("python3 ./LolexToolsInstaller.py")
     exit(None)
-if LolexToolsOptions.compiledon<9.0:
+if LolexToolsOptions.compiledon<9.00001:
     if LolexToolsOptions.compiledon<8.3:
         shutil.copy("./update/8.3release.py","./")
         if system == "Windows":
@@ -36,15 +36,25 @@ if LolexToolsOptions.compiledon<9.0:
         else:
             os.system ("python3 ./8.3release.py")
         os.remove("./8.3release.py")
-    shutil.copy("./update/9.0n1.py","./")
+    if LolexToolsOptions.compiledon<9.0:
+        shutil.copy("./update/9.0n1.py","./")
+        if system == "Windows":
+            if sys.version_info.minor>5:
+                os.system ("py ./9.0n1.py")
+            else:
+                os.system("python ./9.0n1.py")
+        else:
+            os.system ("python3 ./9.0n1.py")
+        os.remove("./9.0n1.py")
+    shutil.copy("./update/9.0nann1.py","./")
     if system == "Windows":
         if sys.version_info.minor>5:
-            os.system ("py ./9.0n1.py")
+            os.system ("py ./9.0nann1.py")
         else:
-            os.system("python ./9.0n1.py")
+            os.system("python ./9.0nann1.py")
     else:
-        os.system ("python3 ./9.0n1.py")
-    os.remove("./9.0n1.py")
+        os.system ("python3 ./9.0nann1.py")
+    os.remove("./9.0nann1.py")
     print("Restarting to finish updating...")
     if system == "Windows":
         if sys.version_info.minor>5:
